Fix Docker based unit tests

+++ contrib/storage-splunk/src/test/java/org/apache/drill/exec/store/splunk/SplunkTestSuite.java
@@ -28,12 +28,12 @@ import org.apache.drill.test.ClusterFixtureBuilder;
 import org.apache.drill.test.ClusterTest;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.ClassRule;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.testcontainers.DockerClientFactory;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.utility.DockerImageName;
 
@@ -42,6 +42,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.apache.drill.exec.rpc.user.security.testing.UserAuthenticatorTestImpl.TEST_USER_1;
 import static org.apache.drill.exec.rpc.user.security.testing.UserAuthenticatorTestImpl.TEST_USER_2;
+import static org.junit.Assume.assumeTrue;
 
 
 @RunWith(Suite.class)
@@ -107,23 +108,14 @@ public class SplunkTestSuite extends ClusterTest {
     }
   }
 
-  @ClassRule
-  public static GenericContainer<?> splunk = new GenericContainer<>(
-    DockerImageName.parse("splunk/splunk:9.3")
-  )
-    .withExposedPorts(8089, 8089)
-    .withEnv("SPLUNK_START_ARGS", "--accept-license")
-    .withEnv("SPLUNK_PASSWORD", SPLUNK_PASS)
-    .withEnv("SPLUNKD_SSL_ENABLE", "false")
-    .withCopyFileToContainer(
-      org.testcontainers.utility.MountableFile.forHostPath(
-        createDefaultYmlFile().toPath()
-      ),
-      "/tmp/defaults/default.yml"
-    );
+  private static GenericContainer<?> splunk;
 
   @BeforeClass
   public static void initSplunk() throws Exception {
+    assumeTrue(
+      "Docker is not available, skipping container tests",
+      DockerClientFactory.instance().isDockerAvailable()
+    );
     synchronized (SplunkTestSuite.class) {
       if (initCount.get() == 0) {
         ClusterFixtureBuilder builder = new ClusterFixtureBuilder(dirTestWatcher)
@@ -132,6 +124,19 @@ public class SplunkTestSuite extends ClusterTest {
           .configProperty(ExecConstants.IMPERSONATION_ENABLED, true);
         startCluster(builder);
 
+        splunk = new GenericContainer<>(
+          DockerImageName.parse("splunk/splunk:9.3")
+        )
+          .withExposedPorts(8089, 8089)
+          .withEnv("SPLUNK_START_ARGS", "--accept-license")
+          .withEnv("SPLUNK_PASSWORD", SPLUNK_PASS)
+          .withEnv("SPLUNKD_SSL_ENABLE", "false")
+          .withCopyFileToContainer(
+            org.testcontainers.utility.MountableFile.forHostPath(
+              createDefaultYmlFile().toPath()
+            ),
+            "/tmp/defaults/default.yml"
+          );
         splunk.start();
 
         // Wait for Splunk to start and apply configuration from default.yml
@@ -203,6 +208,9 @@ public class SplunkTestSuite extends ClusterTest {
    * This should be called between test classes to prevent disk space exhaustion.
    */
   public static void cleanDispatchDirectory() {
+    if (splunk == null) {
+      return;
+    }
     try {
       logger.info("Cleaning up Splunk dispatch directory...");
       splunk.execInContainer("sh", "-c", "rm -rf /opt/splunk/var/run/splunk/dispatch/*");
@@ -215,7 +223,7 @@ public class SplunkTestSuite extends ClusterTest {
   @AfterClass
   public static void tearDownCluster() {
     synchronized (SplunkTestSuite.class) {
-      if (initCount.decrementAndGet() == 0) {
+      if (initCount.decrementAndGet() == 0 && splunk != null) {
         // Clean up Splunk dispatch files to free disk space before shutdown
         cleanDispatchDirectory();
         splunk.close();
